Split the JavaScript into multiple files. Also plan on making a full automator for registering the commands the bot needs.
