+++ index.js
@@ -1,122 +1,10 @@
-const Request = require('request-promise')
+const Discord = require('discord.js');
 require('dotenv').config();
 
-const Weekdays = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday'] // Used to fetch the index for a day
-
-const ReqHeaders = {
-  'Authorization': 'basic T64Mdy7m[',
-  'Content-Type': 'application/json; charset=utf-8',
-  'credentials': 'include',
-  'Referer': 'https://opentimetable.dcu.ie/',
-  'Origin': 'https://opentimetable.dcu.ie/'
-}
-
-//
-
-function StartOfWeek(DateToFetch) {
-  var CurrentDate = DateToFetch
-  var DateDifference = CurrentDate.getDate() - CurrentDate.getDay() + (CurrentDate.getDay() === 0 ? -6 : 1);
-
-  FirstDayInWeek = new Date(CurrentDate.setDate(DateDifference)).toISOString() // Convert our date to ISOString
-  return FirstDayInWeek.slice(0, -14).concat('T00:00:00.000Z') // Slice the date and add a time for midnight to the end
-  // Outputs: YYYY-MM-DDT00:00:00.000Z
-}
-
-function ConstructRequestBody(Day, DateToFetch, CourseIdentity) {
-  let RequestBodyTemplate = require('./body.json')
-
-  FinalDayNumber = Weekdays.indexOf(Day) + 1
-
-  RequestBodyTemplate['ViewOptions']['Weeks'][0]['FirstDayInWeek'] = StartOfWeek(DateToFetch)
-  RequestBodyTemplate['ViewOptions']['Days'][0]['Name'] = Day
-  RequestBodyTemplate['ViewOptions']['Days'][0]['DayOfWeek'] = Weekdays.indexOf(Day) + 1
-
-  RequestBodyTemplate['CategoryIdentities'][0] = CourseIdentity
-
-  return RequestBodyTemplate
-}
-
-async function FetchCourseCodeIdentity(Query) {
-
-  var ReqPayload = {
-    method: 'POST',
-    uri: `https://opentimetable.dcu.ie/broker/api/CategoryTypes/241e4d36-60e0-49f8-b27e-99416745d98d/Categories/Filter?pageNumber=1&query=${Query}`,
-    headers: ReqHeaders,
-    json: true
-  };
-
-  return new Promise(function (resolve, reject) {
-    Request(ReqPayload) // Send the HTTP Request
-      .then(function (res_body) {
-        let Results = res_body['Results']
-
-        if (Results.length == 0) {
-          reject('Course identity not found with supplied course code.')
-        } else {
-          resolve(res_body['Results'][0]['Identity'])
-        }
-      })
-      .catch(function (err) { // Catch any errors
-        reject(err)
-      });
-  })
-}
-
-async function FetchRawTimetableData(CourseCode, Day, DateToFetch = new Date()) {
-
-  let CourseIdentity = await FetchCourseCodeIdentity(CourseCode);
-
-  return new Promise(function (resolve, reject) {
-    const ReqPayload = {
-      method: 'POST',
-      uri: 'https://opentimetable.dcu.ie/broker/api/categoryTypes/241e4d36-60e0-49f8-b27e-99416745d98d/categories/events/filter',
-      headers: ReqHeaders,
-      body: ConstructRequestBody(Day, DateToFetch, CourseIdentity),
-      json: true
-    };
-
-    Request(ReqPayload) // Send the HTTP Request
-      .then(async function (res_body) {
-        await Promise.all(res_body[0].CategoryEvents.map(async event => {
-          let moduleName = await FetchModuleNameFromCode(event.Name.slice(0, 5));
-          return event.Name = moduleName;
-        }))
-        resolve(res_body)
-      })
-      .catch(function (err) { // Catch any errors
-        reject(err)
-      });
-  })
-}
-
-async function FetchModuleNameFromCode(Query) {
-
-  var ReqPayload = {
-    method: 'POST',
-    uri: `https://opentimetable.dcu.ie/broker/api/CategoryTypes/525fe79b-73c3-4b5c-8186-83c652b3adcc/Categories/Filter?pageNumber=1&query=${Query}`,
-    headers: ReqHeaders,
-    json: true
-  };
-
-  return new Promise(function (resolve, reject) {
-    Request(ReqPayload) // Send the HTTP Request
-      .then(function (res_body) {
-        let Results = res_body['Results'];
-
-        if (Results.length == 0) {
-          reject('Course identity not found with supplied course code.');
-        } else {
-          resolve(res_body['Results'][0]['Name']);
-        }
-      })
-      .catch(function (err) { // Catch any errors
-        reject(err);
-      });
-  });
-}
-
-const Discord = require('discord.js');
 const scheduler = require('node-schedule');
+
+const Labfree = require('./labfree.js')
+const Timetable = require('./timetable.js')
 
 const client = new Discord.Client({ intents: [Discord.GatewayIntentBits.DirectMessages] });
 
@@ -150,7 +38,17 @@ client.on('interactionCreate', async interaction => {
   const { commandName } = interaction;
 
   if (commandName === 'ping') {
-    await interaction.reply('Pong!');
+    await interaction.reply(
+      'Pong!'
+    );
+  }
+
+  if (commandName === 'checkfree') {
+    const reply = Labfree.checkfree(interaction)
+
+    await interaction.reply(
+      reply
+    );
   }
 
   else if (commandName === 'timetable') {
@@ -161,7 +59,7 @@ client.on('interactionCreate', async interaction => {
 
     const shortDay = ['mon', 'tue', 'wed', 'thu', 'fri']
     const longDay = ['monday', 'tuesday', 'wednesday', 'thursday', 'friday']
-    let day = Weekdays[(new Date().getDay() - 1)];
+    let day = Timetable.Weekdays[(new Date().getDay() - 1)];
 
     if (interaction.options.getString('day') || interaction.options.getString('course').split(' ')[1]) {
       day = interaction.options.getString('day') || interaction.options.getString('course').split(' ')[1];
@@ -177,7 +75,7 @@ client.on('interactionCreate', async interaction => {
       }
     }
     
-    FetchRawTimetableData(courseCode, day, new Date())
+    Timetable.FetchRawTimetableData(courseCode, day, new Date())
       .then(async (res) => {
         console.log(res)
         res = res[0].CategoryEvents;
@@ -219,8 +117,8 @@ client.on('interactionCreate', async interaction => {
  * @param {String} course
  */
 const morningUpdate = function (user, course) {
-  const day = Weekdays[(new Date().getDay() - 1)];
-  FetchRawTimetableData(course, day, new Date())
+  const day = Timetable.Weekdays[(new Date().getDay() - 1)];
+  Timetable.FetchRawTimetableData(course, day, new Date())
     .then(async (res) => {
       res = res[0].CategoryEvents;
 
@@ -255,3 +153,5 @@ const morningUpdate = function (user, course) {
 }
 
 client.login(process.env.TOKEN);
+
+// Exporting the timetable stuff might be a good idea. Fuck, splitting them would be good.
